requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
google-generativeai
notion-client
line-bot-sdk
pydantic
pytest